add confirm for removing items fix episode number display

+++ src/components/Home/WatchingEpisodes.js
@@ -31,6 +31,7 @@ let searchAnimeQuery = `
 
 function WatchingEpisodes() {
   const [data, setData] = useState([]);
+  const [confirmRemove, setConfirmRemove] = useState([]);
   const [loading, setLoading] = useState(true);
 
   useEffect(() => {
@@ -81,19 +82,38 @@ function WatchingEpisodes() {
       });
     }
     setData(apiRes);
+    setConfirmRemove(apiRes.map(() => false));
     setLoading(false);
   }
 
-  function removeAnime(index) {
-    let lsData = localStorage.getItem("Animes");
-    lsData = JSON.parse(lsData);
-    lsData.Names.splice(index, 1);
-    lsData = JSON.stringify(lsData);
-    localStorage.setItem("Animes", lsData);
-    // data.splice(index, 1);
-    setData(data => [
-      ...data.slice(0, index),
-      ...data.slice(index + 1),
+  function removeAnime(ev) {
+    if (!confirmRemove.length) return;
+    const index = parseInt(ev.currentTarget.dataset.index);
+    if (confirmRemove[index]) {
+      let lsData = localStorage.getItem("Animes");
+      lsData = JSON.parse(lsData);
+      lsData.Names.splice(index, 1);
+      lsData = JSON.stringify(lsData);
+      localStorage.setItem("Animes", lsData);
+      // data.splice(index, 1);
+      setData((data) => [...data.slice(0, index), ...data.slice(index + 1)]);
+      setConfirmRemove([...confirmRemove.slice(0, index), ...confirmRemove.slice(index + 1)]);
+    } else {
+      setConfirmRemove([
+        ...confirmRemove.slice(0, index),
+        true,
+        ...confirmRemove.slice(index + 1),
+      ]);
+    }
+  }
+
+  function cancelRemoveAnime(ev) {
+    if (!confirmRemove.length) return;
+    const index = parseInt(ev.currentTarget.dataset.index);
+    setConfirmRemove([
+      ...confirmRemove.slice(0, index),
+      false,
+      ...confirmRemove.slice(index + 1),
     ]);
   }
 
@@ -144,14 +164,18 @@ function WatchingEpisodes() {
                     },
                   }}
                 >
-                  <button
-                    className="closeButton"
-                    onClick={() => {
-                      removeAnime(i);
-                    }}
-                  >
-                    <IoClose />
-                  </button>
+                  <ConfirmRemove>
+                      <button
+                        className={"removeButton"+(confirmRemove[i]?" confirm":"")}
+                        data-index={i}
+                        onClick={removeAnime}
+                      >
+                        {confirmRemove[i] ? <span>Remove</span> : <IoClose />}
+                      </button>
+                      {confirmRemove[i] && <button data-index={i} onClick={cancelRemoveAnime}>
+                        Cancel
+                      </button>}
+                  </ConfirmRemove>
                 </IconContext.Provider>
 
                 <Link to={"watch/" + item.link}>
@@ -159,7 +183,7 @@ function WatchingEpisodes() {
                 </Link>
                 <p>{item.name}</p>
                 <p className="episodeNumber">
-                  {"Episode - " + item.episodeNum}
+                  {`Episode ${item.episodeNum}`}
                 </p>
               </Wrapper>
             </SwiperSlide>
@@ -170,17 +194,30 @@ function WatchingEpisodes() {
   );
 }
 
+const ConfirmRemove = styled.div`
+  position: absolute;
+  top: 0;
+`;
+
 const Wrapper = styled.div`
   position: relative;
-
-  .closeButton {
-    position: absolute;
+  button {
     cursor: pointer;
     outline: none;
     border: none;
+    color: #fff;
     padding: 0.5rem;
+    border-radius: 0.2rem;
     background-color: rgba(0, 0, 0, 0.7);
-    border-radius: 0.5rem 0 0.2rem 0;
+    &+button {
+      margin-left: 0.5rem;
+    }
+  }
+  .removeButton {
+    border-top-left-radius: 0.5rem;
+    &.confirm {
+      background-color: rgba(150, 0, 0, 0.7);
+    }
   }
   img {
     width: 160px;
@@ -199,7 +236,7 @@ const Wrapper = styled.div`
   }
 
   p {
-    color: #FFFFFF;
+    color: #ffffff;
     font-size: 1rem;
     font-family: "Gilroy-Medium", sans-serif;
     @media screen and (max-width: 600px) {
@@ -212,7 +249,8 @@ const Wrapper = styled.div`
 
   .episodeNumber {
     font-family: "Gilroy-Regular", sans-serif;
-    color: #23272A;
+    color: #969696;
+    font-size: 0.8em;
   }
 `;
